+++ src/plugin.hpp
@@ -235,3 +235,4 @@ extern Model* modelPads64;
 extern Model* modelXY64;
 extern Model* modelRhythm64;
 extern Model* modelDrums64;
+extern Model* modelObjects64;